+++ .gitignore
@@ -1,2 +1,4 @@
 /node_modules
 /dist
+/.idea
+.DS_Store
