+++ src/index.html
@@ -46,7 +46,8 @@
                     <input class="popup__input popup__input_type_name" id='name' name="name" placeholder="Имя" required
                            type="text">
                     <span class='error-message' id='error-name'></span>
-                    <input class="popup__input popup__input_type_link-url" id='job' name="job" placeholder="О себе" required
+                    <input class="popup__input popup__input_type_link-url" id='job' name="job" placeholder="О себе"
+                           required
                            type="text">
                     <span class='error-message' id='error-job'></span>
                     <button class="button popup__button" type>Сохранить</button>
