+++ src/userId.js
@@ -1,0 +1,1 @@
+export let userId = "";
